# backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "윤정 법률사무소 API",
    "version": "1.0.0",
    "description": "윤정 법률사무소 홈페이지 백엔드 REST API. 모든 응답은 { data, error, meta } 형식입니다."
  },
  "servers": [
    { "url": "/api/sb", "description": "기본 API 서버" }
  ],
...
  "paths": {
...
    "/blog": {
      "get": {
        "tags": ["Blog"],
        "summary": "블로그 게시글 목록 조회 (공개)",
        "description": "공개된 게시글 목록을 페이지네이션, 카테고리 필터와 함께 반환합니다.",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 }, "description": "페이지 번호" },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 }, "description": "페이지당 항목 수" },
          { "name": "category", "in": "query", "schema": { "type": "string" }, "description": "카테고리 필터" }
        ],
        "responses": {
          "200": {
            "description": "게시글 목록",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/BlogPost" }
                        },
                        "meta": { "$ref": "#/components/schemas/PaginationMeta" }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Blog"],
        "summary": "블로그 게시글 생성 (관리자)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BlogPostCreate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "생성된 게시글",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/BlogPost" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/blog/{slug}": {
      "get": {
        "tags": ["Blog"],
        "summary": "슬러그로 게시글 조회 (공개)",
        "description": "슬러그로 게시글을 조회하고 조회수를 증가시킵니다.",
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" }, "description": "게시글 슬러그" }
        ],
        "responses": {
          "200": {
            "description": "게시글 상세",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/BlogPost" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/blog/{id}": {
      "patch": {
        "tags": ["Blog"],
        "summary": "게시글 수정 (관리자)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "게시글 ID (UUID)" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BlogPostCreate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "수정된 게시글",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/BlogPost" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Blog"],
        "summary": "게시글 삭제 (관리자)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "게시글 ID (UUID)" }
        ],
        "responses": {
          "200": {
            "description": "삭제 결과",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/portal/cases": {
      "get": {
        "tags": ["Portal"],
        "summary": "내 사건 목록 조회",
        "security": [{ "portalAuth": [] }],
        "responses": {
          "200": {
            "description": "사건 목록",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/CaseFile" }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/portal/cases/{id}": {
      "get": {
        "tags": ["Portal"],
        "summary": "사건 상세 조회 (문서 + 최근 메시지 포함)",
        "description": "본인 사건만 조회 가능합니다. 관련 문서와 최근 메시지 10건을 포함합니다.",
        "security": [{ "portalAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" }, "description": "사건 ID" }
        ],
        "responses": {
          "200": {
            "description": "사건 상세",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "allOf": [
                            { "$ref": "#/components/schemas/CaseFile" },
                            {
                              "type": "object",
                              "properties": {
                                "documents": { "type": "array", "items": { "type": "object" } },
                                "messages": { "type": "array", "items": { "$ref": "#/components/schemas/CaseMessage" } }
                              }
                            }
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/portal/admin/cases": {
      "get": {
        "tags": ["Portal Admin"],
        "summary": "전체 사건 목록 (관리자)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 }, "description": "페이지 번호" },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 }, "description": "페이지당 항목 수 (최대 100)" }
        ],
        "responses": {
          "200": {
            "description": "사건 목록",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/CaseFile" }
                        },
                        "meta": { "$ref": "#/components/schemas/PaginationMeta" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["Portal Admin"],
        "summary": "사건 생성 (관리자)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["clientId", "title"],
                "properties": {
                  "clientId": { "type": "string", "format": "uuid", "description": "고객 ID" },
                  "title": { "type": "string", "description": "사건 제목" },
                  "lawyerId": { "type": "string", "nullable": true, "description": "담당 변호사 ID" },
                  "description": { "type": "string", "nullable": true, "description": "사건 설명" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "생성된 사건",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/CaseFile" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/portal/admin/cases/{id}": {
      "patch": {
        "tags": ["Portal Admin"],
        "summary": "사건 수정 (관리자)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" }, "description": "사건 ID" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": { "type": "string", "description": "사건 상태" },
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "lawyerId": { "type": "string", "nullable": true }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "수정된 사건",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/CaseFile" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "관리자 인증 토큰 (POST /admin-users/login에서 발급)"
      },
      "portalAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-portal-token",
        "description": "포털 세션 토큰 (POST /portal/login에서 발급, TTL 24시간)"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "description": "모든 API의 공통 응답 형식",
        "properties": {
          "data": { "description": "응답 데이터 (성공 시)", "nullable": true },
          "error": { "type": "string", "description": "에러 메시지 (실패 시)", "nullable": true },
          "meta": { "type": "object", "description": "부가 정보 (페이지네이션 등)", "nullable": true }
        }
      },
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "page": { "type": "integer", "description": "현재 페이지" },
          "limit": { "type": "integer", "description": "페이지당 항목 수" },
          "total": { "type": "integer", "description": "전체 항목 수" },
          "totalPages": { "type": "integer", "description": "전체 페이지 수" }
        }
      },
...
      "BlogPost": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "slug": { "type": "string" },
          "content": { "type": "string", "description": "HTML 콘텐츠" },
          "excerpt": { "type": "string", "nullable": true, "description": "요약" },
          "category": { "type": "string", "nullable": true },
          "tags": { "type": "string", "nullable": true, "description": "쉼표 구분 태그" },
          "thumbnailUrl": { "type": "string", "nullable": true },
          "isPublished": { "type": "integer", "description": "공개 여부 (0 또는 1)" },
          "viewCount": { "type": "integer" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "BlogPostCreate": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "slug": { "type": "string" },
          "content": { "type": "string" },
          "excerpt": { "type": "string" },
          "category": { "type": "string" },
          "tags": { "type": "string" },
          "thumbnailUrl": { "type": "string" },
          "isPublished": { "type": "integer" }
        },
        "required": ["title", "content"]
      },
...
      "CaseFile": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "clientId": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "status": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "lawyerId": { "type": "string", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "CaseMessage": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "caseFileId": { "type": "string", "format": "uuid" },
          "senderId": { "type": "string", "nullable": true },
          "senderType": { "type": "string", "enum": ["client", "lawyer"], "description": "발신자 유형" },
          "content": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "잘못된 요청",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                { "$ref": "#/components/schemas/ApiResponse" },
                {
                  "type": "object",
                  "properties": {
                    "data": { "type": "null" },
                    "error": { "type": "string", "example": "필수 항목이 누락되었습니다" }
                  }
                }
              ]
            }
          }
        }
      },
      "Unauthorized": {
        "description": "인증 실패",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                { "$ref": "#/components/schemas/ApiResponse" },
                {
                  "type": "object",
                  "properties": {
                    "data": { "type": "null" },
                    "error": { "type": "string", "example": "인증이 필요합니다" }
                  }
                }
              ]
            }
          }
        }
      },
      "NotFound": {
        "description": "리소스를 찾을 수 없음",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                { "$ref": "#/components/schemas/ApiResponse" },
                {
                  "type": "object",
                  "properties": {
                    "data": { "type": "null" },
                    "error": { "type": "string", "example": "리소스를 찾을 수 없습니다" }
                  }
                }
              ]
            }
          }
        }
      }
    }
  }
}